requirements.txt
streamlit
sentence-transformers
keybert
transformers
pypdf
python-docx
torch
scikit-learn
